remove Result enum
ForthEvaluate.eval() now returns a Bool for homogeneity with other
functions returning a success flag.

Returning a Void? was attempted (() on success, nil on failure) to be in
line with functions returning an optional type and nil on failure.  I
reverted it as it seemed less intuitive and assigning a Void? to a local
variable causes a compile warning.